refactor(session-store): extract persistence owner
